Revert "OrderBookRow use Decimal instead of float"
This reverts commit 1f75e7fcdbd2dc203250c720853a81e27df0ae7a.

+++ hummingbot/core/data_type/order_book_row.py
@@ -8,8 +8,8 @@ class OrderBookRow(namedtuple("_OrderBookRow", "price, amount, update_id")):
     """
     Used to apply changes to OrderBook. OrderBook classes uses float internally for better performance over Decimal.
     """
-    price: Decimal
-    amount: Decimal
+    price: float
+    amount: float
     update_id: int
